// components/runtime/src/ts/generated/internal/AllowedMentions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MentionParseTypes } from "./MentionParseTypes";

export interface AllowedMentions { parse: Array<MentionParseTypes>, users: Array<string>, roles: Array<string>, repliedUser: boolean, }